fixed bug that was giving negative scaled fitness values. This was because the populations coverage was not always updated every time somthing in the population changed. Should be working now.

+++ Comp 524 Project/src/Population.h
@@ -85,7 +85,7 @@ public:
 private:
 	Organism** population;
 	int populationSize;
-	int totalFitness;
+	long long totalFitness;
 	int* edgesCovered;
 	int* predicatesCovered;
 	double coverageRatio;
